Refactor folder structure and namings
