+++ src/resources/apps/users/resource_users.py
@@ -1,5 +1,4 @@
 import shortuuid
-
 from flask_restful import reqparse
 from rubix_http.exceptions.exception import NotFoundException, UnauthorizedException, BadDataException
 from rubix_http.resource import RubixResource
