+++ src/routes.py
@@ -2,15 +2,15 @@ from flask import Blueprint
 from flask_restful import Api
 
 from src.resources.admin.resource_fcm_server import FcmServerResource
-from src.resources.admin.resource_site import SiteResourceList, SiteResourceByUUID
+from src.resources.admin.resource_sites import AdminSitesResourceList, AdminSitesResourceByUUID
 from src.resources.admin.resource_user_site import UserSiteResourceList, UserSiteResourceByUUID
 from src.resources.admin.resource_users import *
-from src.resources.third_party.resource_device import DeviceResourceList, DeviceResourceByUUID, \
-    DeviceResourceByDeviceId
-from src.resources.third_party.resource_mqtt import MqttTopicsResource, MqttConfigResource
-from src.resources.third_party.resource_own_user_sites import OwnUserSitesResourceList
-from src.resources.third_party.resource_user import UserResource
-from src.resources.third_party.resource_users import *
+from src.resources.apps.own_users.resource_devices import DevicesResourceList, DevicesResourceByUUID, \
+    DevicesResourceByDeviceId
+from src.resources.apps.own_users.resource_mqtt import MqttTopicsResource, MqttConfigResource
+from src.resources.apps.own_users.resource_sites import SitesResourceList
+from src.resources.apps.own_users.resource_user import UserResource
+from src.resources.apps.users.resource_users import *
 from src.system.resources.ping import Ping
 
 # Admin
@@ -40,18 +40,18 @@ api_users.add_resource(UserSiteResourceByUUID, '/sites/uuid/<string:uuid>')
 api_apps_users = Api(bp_users)
 api_apps_users.add_resource(UsersCreateResource, '', endpoint="create")
 api_apps_users.add_resource(UsersLoginResource, '/login', endpoint="login")
-api_apps_users.add_resource(UsersChangePasswordResource, '/change_password')
 api_apps_users.add_resource(UsersCheckByUsernameResource, '/check/username', endpoint='check_username')
 api_apps_users.add_resource(UsersCheckByEmailResource, '/check/email', endpoint='check_email')
-api_apps_users.add_resource(UsersRefreshToken, '/refresh_token')
 
 # 4 => Apps
 api_apps_own_users = Api(bp_own_users)
 api_apps_own_users.add_resource(UserResource, '')
-api_apps_own_users.add_resource(DeviceResourceList, '/devices')
-api_apps_own_users.add_resource(DeviceResourceByUUID, '/devices/uuid/<string:uuid>')
-api_apps_own_users.add_resource(DeviceResourceByDeviceId, '/devices/device_id/<string:device_id>')
-api_apps_own_users.add_resource(OwnUserSitesResourceList, '/sites')
+api_apps_own_users.add_resource(UsersChangePasswordResource, '/change_password')
+api_apps_own_users.add_resource(UsersRefreshToken, '/refresh_token')
+api_apps_own_users.add_resource(DevicesResourceList, '/devices')
+api_apps_own_users.add_resource(DevicesResourceByUUID, '/devices/uuid/<string:uuid>')
+api_apps_own_users.add_resource(DevicesResourceByDeviceId, '/devices/device_id/<string:device_id>')
+api_apps_own_users.add_resource(SitesResourceList, '/sites')
 
 # 5 => Apps
 api_apps_configs = Api(bp_configs)
@@ -64,5 +64,5 @@ api_fcm_server.add_resource(FcmServerResource, '')
 
 # 7 => Admin
 api_sites = Api(bp_sites)
-api_sites.add_resource(SiteResourceList, '')
-api_sites.add_resource(SiteResourceByUUID, '/uuid/<string:uuid>')
+api_sites.add_resource(AdminSitesResourceList, '')
+api_sites.add_resource(AdminSitesResourceByUUID, '/uuid/<string:uuid>')
